+++ app/src/main/java/com/cryptopaths/cryptofm/tasks/EncryptTask.java
@@ -138,7 +138,8 @@ public class EncryptTask extends AsyncTask<Void,String,String> {
     @Override
     protected void onCancelled() {
         for (File f : mCreatedFiles) {
-            if(f.getAbsolutePath().contains(SharedData.EXTERNAL_SDCARD_ROOT_PATH)){
+            if(SharedData.EXTERNAL_SDCARD_ROOT_PATH!=null &&
+                    f.getAbsolutePath().contains(SharedData.EXTERNAL_SDCARD_ROOT_PATH)){
                 FileDocumentUtils.getDocumentFile(f).delete();
             }else{
                 f.delete();
